+++ src/order/order.go
@@ -7,7 +7,7 @@ type PurchaseOrder struct {
 	ValuableID      string                  `json:"stock_id"`
 	PortfolioID     string                  `json:"portfolio"`
 	ExchangeID      string                  `json:"exchange"`
-	Amount          float64                 `json:"amount"`
+	Amount          int64                 `json:"amount"`
 	ResponseChannel chan *PurchasedResponse `json:"-"`
 }
 
@@ -18,7 +18,7 @@ type PurchasedResponse struct {
 }
 
 // note this does not validate if the stock exists or not, that's done in the trade() function
-func BuildPurchaseOrder(valuableID, portfolioUUID string, amount float64) *PurchaseOrder {
+func BuildPurchaseOrder(valuableID, portfolioUUID string, amount int64) *PurchaseOrder {
 	return &PurchaseOrder{
 		//ExchangeID:      exchangeID,
 		ValuableID:      valuableID,
